project/requirements.txt
asgiref==3.8.1
dj-database-url==2.1.0
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.1
psycopg2==2.9.9
PyJWT==1.7.1
sqlparse==0.4.4
typing_extensions==4.11.0
tzdata==2024.1